Fix image issue in markdown
Markdown images were not displaying at all in posts. This was due to the
library `gatsby-remark-images`, which uses inline styles for its
wrappers (grr). The inline style sets the z-index to -1, which when set
removes the images to the page.

The (addmittedly hacky) fix is providing `gatsby-remark-images` with an
addition style in it's config, which sets the z-index to 0.

Also, the posts folder has been moved into the pages folder.