Fix project links
The `projects` and `writing` paths both use the same slat template.
However, `projects` links to external github pages, while `writing`
links to local pages.

This means that `projects` needed to use `<a>` tags, while `writing`
needed to use Gatsby's `Link` component.

So the `Slat` component now accepts a `static` param, that will render
an `<a>` tag if true, and a `Link` component otherwise.